backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
python-dotenv==1.0.0
